Book Management Database
This is a simple CRUD program that I made. It takes the form of a book database system. It allows you to add books to the system, each of which has a title and author. You can also list the books in the database, edit the current information of a book, or delete a book from the system.